% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_check_statement.r
\name{parse_check_statement}
\alias{parse_check_statement}
\title{Parse a Chase bank statement (saved in a PDF format) into a tidyverse-friendly dataframe}
\usage{
parse_check_statement(file_path, report_year)
}
\arguments{
\item{file_path}{A file path which indicates a PDF file that contains Chase bank statement.}

\item{report_year}{The year in which the bank statement was reported. The data type of this argument should be numeric (e.g., 2020).}
}
\value{
A dataframe with eight columns ("Date, "Description", "Amount", "Balance", "Card", "Venmo", "Withdraw", "Deposit")
}
\description{
Parse a Chase bank statement (saved in a PDF format) into a tidyverse-friendly dataframe
}
